Separated Base, Component and DataType Builders to individual packages.
