+++ src/main/java/com/lswebworld/datatypes/DistanceBuilder.java
@@ -1,5 +1,6 @@
 package com.lswebworld.datatypes;
 
+import com.lswebworld.base.AbstractQuantityBuilder;
 import org.hl7.fhir.r4.model.Distance;
 
 public class DistanceBuilder extends AbstractQuantityBuilder<Distance> {
