+++ src/main/java/com/lswebworld/datatypes/DurationBuilder.java
@@ -1,5 +1,6 @@
 package com.lswebworld.datatypes;
 
+import com.lswebworld.base.AbstractQuantityBuilder;
 import org.hl7.fhir.r4.model.Duration;
 
 public class DurationBuilder extends AbstractQuantityBuilder<Duration> {
