+++ src/main/java/com/lswebworld/datatypes/MoneyBuilder.java
@@ -1,5 +1,6 @@
 package com.lswebworld.datatypes;
 
+import com.lswebworld.base.AbstractElementBuilder;
 import org.hl7.fhir.r4.model.Money;
 
 public class MoneyBuilder extends AbstractElementBuilder<Money, MoneyBuilder> {
