+++ src/main/java/com/lswebworld/datatypes/MoneyQuantityBuilder.java
@@ -1,5 +1,6 @@
 package com.lswebworld.datatypes;
 
+import com.lswebworld.base.AbstractQuantityBuilder;
 import org.hl7.fhir.r4.model.MoneyQuantity;
 
 public class MoneyQuantityBuilder extends AbstractQuantityBuilder<MoneyQuantity> {
