+++ src/main/java/com/lswebworld/datatypes/PeriodBuilder.java
@@ -1,5 +1,6 @@
 package com.lswebworld.datatypes;
 
+import com.lswebworld.base.AbstractElementBuilder;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.util.Date;
