+++ src/main/java/com/lswebworld/datatypes/RangeBuilder.java
@@ -1,5 +1,6 @@
 package com.lswebworld.datatypes;
 
+import com.lswebworld.base.AbstractElementBuilder;
 import org.hl7.fhir.r4.model.Range;
 import org.hl7.fhir.r4.model.SimpleQuantity;
